refactor(api-gen): consolidate member-specific transforms
Refactoring towards eliminating catch-all directories like "helpers"
